Experience level up; experience distribute among exposed; fix tests

+++ mon-gen/src/base/types.rs
@@ -114,7 +114,7 @@ pub mod monster
 	pub type CaughtId = u8;
 
 	/// A experience type for Monster.
-	pub type ExperienceType = u8;
+	pub type ExperienceType = u32;
 }
 
 pub mod battle
